Move all authentication-related views, forms, and tests into the 'users' app.
As the authentication backend grows more complex, it makes less and less sense to combine it with the 'dashboard' app, which is largely used for controlling the internal UI after users have logged in. For that reason, this commit is completing the move of all authentication functionality into the 'users' app, which is much more closely tied to authentication already.